Test that we only trigger automatic rolling updates when we want to.
